requirements/base.txt
Django>=5.2
python-decouple
dj-database-url
whitenoise
gunicorn
psycopg2-binary
django-ckeditor